Remove not needed in frontend
