+++ frontend/package.json
@@ -13,9 +13,7 @@
         "@headlessui/react": "^2.0.4",
         "dompurify": "^3.0.9",
         "fast-deep-equal": "^3.1.3",
-        "highlight.js": "11.8.0",
         "jwt-decode": "^4.0.0",
-        "marked": "6.0.0",
         "next": "14.2.2",
         "react": "^18",
         "react-dom": "^18",
